Refactor: move BucketOwnerCondition static method to S3Owner
