+++ hadoop-ozone/s3gateway/src/main/java/org/apache/hadoop/ozone/s3/endpoint/EndpointBase.java
@@ -217,7 +217,7 @@ public abstract class EndpointBase implements Auditor {
       throws IOException, OS3Exception {
     try {
       OzoneBucket bucket = getBucket(s3BucketName);
-      BucketOwnerCondition.verify(headers, bucket.getOwner());
+      S3Owner.verify(headers, bucket.getOwner());
       client.getObjectStore().deleteS3Bucket(s3BucketName);
     } catch (OMException ex) {
       if (ex.getResult() == ResultCodes.PERMISSION_DENIED) {
